+++ dueca/CommObjectElementReaderBase.hxx
@@ -69,6 +69,7 @@ struct elementdata
   typedef void* elt_key_type;
 };
 
+
 template<typename T>
 struct elementdata<dco_read_iterable,T>
 {
@@ -89,6 +90,17 @@ struct elementdata<dco_read_map,T>
   typedef typename T::key_type    elt_key_type;
 };
 
+template<typename T>
+struct elementdata<dco_read_optional,T>
+{
+  const T *object;
+  bool have_read;
+  elementdata<dco_traits<typename T::value_type>, typename T::value_type> value;
+  typedef typename elementdata<dco_traits<typename T::value_type>, typename T::value_type>::value_type value_type;
+  typedef typename elementdata<dco_traits<typename T::value_type>, typename T::value_type>::elt_value_type elt_value_type;
+  typedef typename elementdata<dco_traits<typename T::value_type>, typename T::value_type>::elt_key_type elt_key_type;
+};
+
 
 template<class T>
 class ReadElement: public elementdata< typename dco_traits<T>::rtype, T>,
@@ -106,6 +118,8 @@ private:
   { par::ii = data.begin(); par::object = &data; }
   void constructor(const T& data, const dco_read_map&)
   { par::ii = data.begin(); par::object = &data; }
+  void constructor(const T& data, const dco_read_optional&)
+  { par::object = &data; par::have_read = false; }
 
 public:
   bool isEnd() const final
@@ -118,6 +132,8 @@ private:
   { return par::ii == par::object->end(); }
   inline bool _isEnd(const dco_read_map&) const
   { return par::ii == par::object->end(); }
+  inline bool _isEnd(const dco_read_optional&) const
+  { return par::have_read; }
 
 public:
   MemberArity getArity() const
@@ -130,6 +146,8 @@ private:
   { return Iterable; }
   inline MemberArity getArity(const dco_read_map&) const
   { return Mapped; }
+  inline MemberArity getArity(const dco_read_optional&) const
+  { return Single; }
 
 public:
   size_t size() const
@@ -142,6 +160,8 @@ private:
   { return par::object->size(); }
   inline size_t size(const dco_read_map&) const
   { return par::object->size(); }
+  inline size_t size(const dco_read_optional&) const
+  { return size_t(1); }
 
 public:
   bool isNested() const
@@ -157,6 +177,7 @@ private:
   inline void step(const dco_read_single&) { par::have_read = true; }
   inline void step(const dco_read_iterable&) { ++par::ii; }
   inline void step(const dco_read_map&) { ++par::ii; }
+  inline void step(const dco_read_optional&) { par::have_read = true; }
 
 public:
   CommObjectReader recurse(std::string& key)
@@ -189,6 +210,9 @@ private:
   const typename par::elt_value_type& get_object(const dco_read_map& x)
   { const typename par::elt_value_type *tmp = &(par::ii->second);
     step(x); return *tmp; }
+  const typename par::elt_value_type& get_object(const dco_read_optional& x)
+  { const typename par::value_type *tmp = &(par::object->value);
+    step(x); return *tmp; }
 
   const typename par::elt_value_type& peek_object(const dco_read_single& x)
   { const typename par::elt_value_type *tmp = par::object;
@@ -199,15 +223,22 @@ private:
   const typename par::elt_value_type& peek_object(const dco_read_map& x)
   { const typename par::elt_value_type *tmp = &(par::ii->second);
     return *tmp; }
+  const typename par::elt_value_type& peek_object(const dco_read_optional& x)
+  { const typename par::value_type *tmp = &(par::object->value);
+    return *tmp; }
 
   void get_key(std::string&, const dco_read_single&) { }
   void get_key(std::string&, const dco_read_iterable&) { }
   void get_key(std::string& key, const dco_read_map&)
   { key = boost::any_cast<std::string>(par::ii->first); }
+  void get_key(std::string& key, const dco_read_optional&) 
+  { key = boost::lexical_cast<std::string>(par::object->valid); } 
   void get_key(boost::any&, const dco_read_single&) { }
   void get_key(boost::any&, const dco_read_iterable&) { }
   void get_key(boost::any& key, const dco_read_map&)
   { key = par::ii->first; }
+  void get_key(boost::any& key, const dco_read_optional&) 
+  { key = par::object->value; } 
 
 public:
   void read(std::string& res, std::string& key)