+++ dueca/dcoprint.cxx
@@ -1,0 +1,25 @@
+/* ------------------------------------------------------------------   */
+/*      item            : dcoprint.cxx
+        made by         : Rene' van Paassen
+        date            : 230614
+        category        : body file
+        description     :
+        changes         : 230614 first version
+        language        : C++
+        copyright       : (c) 23 TUDelft-AE-C&S
+*/
+
+#include "dcoprint.hxx"
+
+
+dcoprint::dcoprint()
+{
+
+}
+
+
+dcoprint::~dcoprint()
+{
+
+}
+