+++ dueca/varvector.hxx
@@ -23,6 +23,7 @@
 #include <algorithm>
 #include <inttypes.h>
 #include <type_traits>
+#include <sstream>
 
 DUECA_NS_START;
 
@@ -256,14 +257,22 @@ public:
 
 /** Helper, for DCO object handling */
 template <typename D>
-struct dco_traits<varvector<D> > : dco_traits_iterable { };
-
-/** Helper, for DCO object handling */
-template <typename D>
-struct pack_traits<varvector<D> >: public pack_var_size, unpack_resize { };
-/** Helper, for DCO object handling */
-template <typename D>
-struct diffpack_traits<varvector<D> >: public diffpack_vector { };
+struct dco_traits<varvector<D> > : dco_traits_iterable,
+  pack_var_size, unpack_resize, diffpack_vector
+{ 
+  /** Representative name */
+  static const char* getclassname()
+  {
+    static std::stringstream cname;
+    if (cname.str().size() == 0) {
+      cname << "varvector<" 
+            << dco_traits<D>::getclassname() << ">";
+    }
+    return cname.str().c_str();
+  }
+  /** Value type for the elements of a trait's target */
+  typedef D value_type;
+};
 
 DUECA_NS_END;